move lrucache to borg.algorithms
