+++ src/borg/testsuite/lrucache.py
@@ -2,7 +2,7 @@ from tempfile import TemporaryFile
 
 import pytest
 
-from ..lrucache import LRUCache
+from ..algorithms.lrucache import LRUCache
 
 
 class TestLRUCache:
